+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Monochrome icon for Android 13 and higher
+- The gallery now scrolls as you swipe content in the viewer. 
+Once you returned, the last viewed media is visible in the list
+- Turkish translation
+
 ## [1.8.1] - 2023-05-28
 
 ### Fixed